REST Classes Added + Comments Added
+ REST_API.cs - This will handle all of our REST (GET, POST, etc.)
calls. Added so far is the UPC look up from scanned items and the
ItemName look up for the Recipes from Spoonacular.
+ SpoonacularClasses.cs - Classes to hold our Deserialized Json values.
+ UPCItemDBClasses.cs - Classes to hold our Deserialized Json values.

==Additions==
+ BarcodeScanner.cs - Added upcJson as our variable to hold our
deserialized json. We use this to grab our item name, which will need
to be fixed. We then use that name to grab our recipe. Still a WIP.
+ MainActivity.cs - Added testing functionality so that instead of
scanning we can hit this button. Uncomment the ones commented to test
the upc GET request.