Batch Delete Favorites
We can now batch delete items from the favorites list. However, there
is a collection (ObservableCollection) that allows for changes to made
and then automatically updated rather than what we are doing now. What
we are doing now seems very inefficient, but I have spent the past 3
days trying to actually find a way to do this, so this will have to
suffice for now.