attempt to allow persistant login on app close
unfortunatly these changes do not fix the problem. i am unable to read
the application output to debug what is wrong with my code. i am pretty
sure an exception is being thrown in the try in Auth0ClientLocal on
line 25, but i'm not sure what exception is being thrown and therefore
am having a difficult time debugging